Stock Market/requirements.txt
openai
python-dotenv
yfinance
packaging
uvicorn
groq
